[Refactor] ItemService 리펙토링
